refactored all error inducing parts to fit the app router
