+++ package.json
@@ -24,7 +24,7 @@
     "axios": "^1.7.7",
     "framer-motion": "11.0.24",
     "next": "14.1.4",
-    "next-intl": "3.10.0",
+    "next-intl": "^3.26.3",
     "next-plausible": "3.12.0",
     "next-sanity": "8.5.0",
     "next-seo": "6.5.0",
