+++ src/app/events/client-page.tsx
@@ -1,0 +1,29 @@
+"use client";
+
+import { useTranslations } from 'next-intl'
+import { LargeEventGrid } from '~/components/event-components/LargeEventGrid'
+import { EventTile } from '~/components/event-components/EventTile'
+import Layout from '~/components/Layout'
+
+export const EventsPageContent = ({ currentEvents, previousEvents }) => {
+  const t = useTranslations('Events')
+
+  return (
+    <Layout>
+      <div className="w-screen justify-items-center sm:w-3/4 mx-auto">
+        <h1 className="text-2xl my-8">{t('current')}</h1>
+        {currentEvents.length > 0 ? (
+          <LargeEventGrid currentEvents={currentEvents} />
+        ) : (
+          <h1 className="text-center text-2xl">{t('none')}</h1>
+        )}
+        <h1 className="text-2xl my-10">{t('past')}</h1>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5 text-sm mb-[10%]">
+          {previousEvents.map((event) => (
+            <EventTile key={event._id} event={event} />
+          ))}
+        </div>
+      </div>
+    </Layout>
+  )
+}
